requirements.txt
flask
pytesseract
pillow
easyocr
opencv-python
torch
torchvision